breakpoint - Got the test_mmap working.
Now the regular test program and the test_mmap program both work.

+++ c/breakpoint/breakpoint.h
@@ -2,11 +2,13 @@
 #ifndef __BREAKPOINT_H__
 #define __BREAKPOINT_H__
 
+#include <stdint.h>
+
 typedef int (*pre_processor_callback)
 	(int arg_01, int arg_02, int arg_03, int arg_04, int arg_05,
 	 int arg_06, int arg_07, int arg_08, int arg_09, int arg_10);
 
-typedef int (*post_processor_callback) (int retcode);
+typedef int (*post_processor_callback) (uint64_t retcode);
 
 int breakpoint_handler_init(void);
